More faff
Adds comments around the insight that... If you enforce MFA interface
you can define things like an equality fn for steps and pipelines which
means you can get nice testing without the need for mocking. You would
be able to assert on the contents of the pipeline without having to run
it. Which is tres dope. You are essentially making an AST at that point?

Which makes me wonder can we use Macros to clean up the interface if it
is bad? Probs can't get more simple than MFA though...

However there are implications for the syntax and overall UX for the lib
if we do that. Certainly looks less idiomatic.

A simpler approach is that steps are behaviours. But then all functions
need to be in a module which is subpar

And finally you have functions. If you have functions then the issue is
you needs ta mock if you don't want to run dem steps. Ya needs to.